Flickering Lights Added
When the Pixar knickknack is turned upside-down, the Luxo Jr. light turns on and off. Similarly, when the Gravity Falls knickknack is turned upside-down, red light turns on and off. The lighting of each knickknack does not affect the other. However, it should be noted that, at this stage, the lights flicker when upside-down. This must be changed.